feat: Update examples to use NaiveDate for holiday checks and add new ymd example
