+++ crates/jp-holidays-lib/examples/cache.rs
@@ -1,3 +1,4 @@
+use chrono::NaiveDate;
 use jp_holidays_lib::{client::Client, error::Error};
 
 // Client::init() は非同期に内閣府から祝日情報を取得するため、
@@ -15,14 +16,20 @@ async fn execute() -> Result<(), Box<dyn std::error::Error>> {
     let client = get_client().await?;
 
     // 祝日を取得
-    let maybe_holiday = client.get_holiday_ymd(1955, 11, 23);
-    match maybe_holiday? {
+    let date = NaiveDate::from_ymd_opt(1955, 11, 23).ok_or("存在しない日付です".to_string())?;
+
+    let maybe_holiday = client.get_holiday(date);
+
+    match maybe_holiday {
         Some(holiday) => println!("1955年 11月 23日 は{}", holiday),
         None => println!("1955年 11月 23日 は祝日ではありません"),
     };
 
     // 祝日かどうか確認
-    let is_holiday = client.is_holiday_ymd(1956, 3, 21)?;
+    let date = NaiveDate::from_ymd_opt(1956, 3, 21).ok_or("存在しない日付です".to_string())?;
+
+    let is_holiday = client.is_holiday(date);
+
     println!(
         "1956 3月 21日 は{}",
         if is_holiday {
