+++ crates/jp-holidays-lib/examples/ymd.rs
@@ -1,25 +1,19 @@
-use chrono::NaiveDate;
 use jp_holidays_lib::client::Client;
 
+/// *_ymd サフィックスの関数を使用すると `chrono::NativeDate` を使用せずに実行できます。
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let client = Client::init().await?;
 
     // 祝日を取得
-    let date = NaiveDate::from_ymd_opt(1955, 11, 23).ok_or("存在しない日付です".to_string())?;
-
-    let maybe_holiday = client.get_holiday(date);
-
-    match maybe_holiday {
+    let maybe_holiday = client.get_holiday_ymd(1955, 11, 23);
+    match maybe_holiday? {
         Some(holiday) => println!("1955年 11月 23日 は{}", holiday),
         None => println!("1955年 11月 23日 は祝日ではありません"),
     };
 
     // 祝日かどうか確認
-    let date = NaiveDate::from_ymd_opt(1956, 3, 21).ok_or("存在しない日付です".to_string())?;
-
-    let is_holiday = client.is_holiday(date);
-
+    let is_holiday = client.is_holiday_ymd(1956, 3, 21)?;
     println!(
         "1956 3月 21日 は{}",
         if is_holiday {
